+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ### Added
 
-- Add cluster name label to Cluster API provider Azure (CAPA) Apps and ConfigMaps created with `kubectl-gs template`
+- Add cluster name label to Cluster API provider AWS (CAPA) Apps and ConfigMaps created with `kubectl-gs template`
 
 ## [2.4.0] - 2022-03-21